fix: unimportant compiler warnings
fixes warnings invoked by -Wconversion and -Wshadow

+++ compiler/include/CompilerInst.h
@@ -48,7 +48,7 @@ public:
 
 
     void setBaseThreadCount(const std::string& count) {
-        m_BaseThreadCount = std::stoi(count);
+        m_BaseThreadCount = static_cast<unsigned>(std::stoi(count));
     }
